+++ apps/web/src/app/(app)/claw/kilo-chat/hooks/useMessages.ts
@@ -1,13 +1,20 @@
-export { useMessages, useSendMessage } from '@kilocode/kilo-chat-hooks';
+export {
+  useMessages,
+  useSendMessage,
+  useEditMessage,
+  useDeleteMessage,
+  useAddReaction,
+  useRemoveReaction,
+  useExecuteAction,
+} from '@kilocode/kilo-chat-hooks';
 export type { SendMessageVariables } from '@kilocode/kilo-chat-hooks';
 
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useQueryClient } from '@tanstack/react-query';
 import type { InfiniteData } from '@tanstack/react-query';
 import type { KiloChatClient } from '@kilocode/kilo-chat';
 import type {
   Message,
   ReactionSummary,
-  EditMessageRequest,
   MessageCreatedEvent,
   MessageUpdatedEvent,
   MessageDeletedEvent,
@@ -15,7 +22,6 @@ import type {
   ActionDeliveryFailedEvent,
   ReactionAddedEvent,
   ReactionRemovedEvent,
-  ExecApprovalDecision,
 } from '@kilocode/kilo-chat';
 import { useEffect } from 'react';
 import { kiloclawConversationContext } from '@kilocode/event-service';
@@ -48,259 +54,6 @@ function applyReactionRemoved(
       return { ...r, count: memberIds.length, memberIds };
     })
     .filter(r => r.count > 0);
-}
-
-/**
- * Splice a snapshotted message back into the current cache state. If the
- * message no longer exists in any page (e.g. a concurrent delete event), the
- * cache is left unchanged so we do not resurrect it.
- */
-function restoreMessageInCache(
-  queryClient: ReturnType<typeof useQueryClient>,
-  queryKey: readonly unknown[],
-  snapshot: Message
-): void {
-  queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-    if (!old) return old;
-    let replaced = false;
-    const pages = old.pages.map(page =>
-      page.map(msg => {
-        if (msg.id !== snapshot.id) return msg;
-        replaced = true;
-        return snapshot;
-      })
-    );
-    if (!replaced) return old;
-    return { ...old, pages };
-  });
-}
-
-/**
- * Remove a message by id from the current cache state. Used to roll back the
- * optimistic insert performed by `useSendMessage` without touching any other
- * concurrently-optimistic messages.
- */
-function removeMessageFromCache(
-  queryClient: ReturnType<typeof useQueryClient>,
-  queryKey: readonly unknown[],
-  messageId: string
-): void {
-  queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-    if (!old) return old;
-    return {
-      ...old,
-      pages: old.pages.map(page => page.filter(msg => msg.id !== messageId)),
-    };
-  });
-}
-
-function findMessageInCache(
-  queryClient: ReturnType<typeof useQueryClient>,
-  queryKey: readonly unknown[],
-  messageId: string
-): Message | undefined {
-  const data = queryClient.getQueryData<InfiniteData<Message[]>>(queryKey);
-  if (!data) return undefined;
-  for (const page of data.pages) {
-    const match = page.find(msg => msg.id === messageId);
-    if (match) return match;
-  }
-  return undefined;
-}
-
-export function useEditMessage(client: KiloChatClient, conversationId: string | null) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({ messageId, ...req }: EditMessageRequest & { messageId: string }) =>
-      client.editMessage(messageId, req),
-    onMutate: async variables => {
-      if (!conversationId) return;
-      const queryKey = ['kilo-chat', 'messages', conversationId];
-      await queryClient.cancelQueries({ queryKey });
-      const snapshot = findMessageInCache(queryClient, queryKey, variables.messageId);
-      queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-        if (!old) return old;
-        return {
-          ...old,
-          pages: old.pages.map(page =>
-            page.map(msg =>
-              msg.id === variables.messageId
-                ? { ...msg, content: variables.content, clientUpdatedAt: variables.timestamp }
-                : msg
-            )
-          ),
-        };
-      });
-      return { queryKey, snapshot };
-    },
-    onError: (_err, _variables, context) => {
-      if (!context?.snapshot) return;
-      restoreMessageInCache(queryClient, context.queryKey, context.snapshot);
-    },
-  });
-}
-
-export function useDeleteMessage(client: KiloChatClient, conversationId: string | null) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({ messageId, conversationId }: { messageId: string; conversationId: string }) =>
-      client.deleteMessage(messageId, { conversationId }),
-    onMutate: async variables => {
-      if (!conversationId) return;
-      const queryKey = ['kilo-chat', 'messages', conversationId];
-      await queryClient.cancelQueries({ queryKey });
-      const snapshot = findMessageInCache(queryClient, queryKey, variables.messageId);
-      queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-        if (!old) return old;
-        return {
-          ...old,
-          pages: old.pages.map(page =>
-            page.map(msg => (msg.id === variables.messageId ? { ...msg, deleted: true } : msg))
-          ),
-        };
-      });
-      return { queryKey, snapshot };
-    },
-    onError: (_err, _variables, context) => {
-      if (!context?.snapshot) return;
-      restoreMessageInCache(queryClient, context.queryKey, context.snapshot);
-    },
-  });
-}
-
-export function useAddReaction(
-  client: KiloChatClient,
-  conversationId: string | null,
-  currentUserId: string
-) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({ messageId, emoji }: { messageId: string; emoji: string }) =>
-      client.addReaction(messageId, { conversationId: conversationId ?? '', emoji }),
-    onMutate: async variables => {
-      if (!conversationId) return;
-      const queryKey = ['kilo-chat', 'messages', conversationId];
-      await queryClient.cancelQueries({ queryKey });
-      const snapshot = findMessageInCache(queryClient, queryKey, variables.messageId);
-      queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-        if (!old) return old;
-        return {
-          ...old,
-          pages: old.pages.map(page =>
-            page.map(msg =>
-              msg.id !== variables.messageId
-                ? msg
-                : {
-                    ...msg,
-                    reactions: applyReactionAdded(msg.reactions, variables.emoji, currentUserId),
-                  }
-            )
-          ),
-        };
-      });
-      return { queryKey, snapshot };
-    },
-    onError: (_err, _variables, context) => {
-      if (!context?.snapshot) return;
-      restoreMessageInCache(queryClient, context.queryKey, context.snapshot);
-    },
-  });
-}
-
-export function useRemoveReaction(
-  client: KiloChatClient,
-  conversationId: string | null,
-  currentUserId: string
-) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({ messageId, emoji }: { messageId: string; emoji: string }) =>
-      client.removeReaction(messageId, { conversationId: conversationId ?? '', emoji }),
-    onMutate: async variables => {
-      if (!conversationId) return;
-      const queryKey = ['kilo-chat', 'messages', conversationId];
-      await queryClient.cancelQueries({ queryKey });
-      const snapshot = findMessageInCache(queryClient, queryKey, variables.messageId);
-      queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-        if (!old) return old;
-        return {
-          ...old,
-          pages: old.pages.map(page =>
-            page.map(msg =>
-              msg.id !== variables.messageId
-                ? msg
-                : {
-                    ...msg,
-                    reactions: applyReactionRemoved(msg.reactions, variables.emoji, currentUserId),
-                  }
-            )
-          ),
-        };
-      });
-      return { queryKey, snapshot };
-    },
-    onError: (_err, _variables, context) => {
-      if (!context?.snapshot) return;
-      restoreMessageInCache(queryClient, context.queryKey, context.snapshot);
-    },
-  });
-}
-
-export function useExecuteAction(
-  client: KiloChatClient,
-  conversationId: string | null,
-  currentUserId: string
-) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({
-      messageId,
-      groupId,
-      value,
-    }: {
-      messageId: string;
-      groupId: string;
-      value: ExecApprovalDecision;
-    }) => client.executeAction(conversationId ?? '', messageId, { groupId, value }),
-    onMutate: async variables => {
-      if (!conversationId) return;
-      const queryKey = ['kilo-chat', 'messages', conversationId];
-      await queryClient.cancelQueries({ queryKey });
-      const snapshot = findMessageInCache(queryClient, queryKey, variables.messageId);
-      // Optimistically mark the action as resolved
-      queryClient.setQueryData<InfiniteData<Message[]>>(queryKey, old => {
-        if (!old) return old;
-        return {
-          ...old,
-          pages: old.pages.map(page =>
-            page.map(msg => {
-              if (msg.id !== variables.messageId) return msg;
-              return {
-                ...msg,
-                content: msg.content.map(block => {
-                  if (block.type !== 'actions') return block;
-                  if (block.groupId !== variables.groupId) return block;
-                  return {
-                    ...block,
-                    resolved: {
-                      value: variables.value,
-                      resolvedBy: currentUserId,
-                      resolvedAt: Date.now(),
-                    },
-                  };
-                }),
-              };
-            })
-          ),
-        };
-      });
-      return { queryKey, snapshot };
-    },
-    onError: (_err, _variables, context) => {
-      if (!context?.snapshot) return;
-      restoreMessageInCache(queryClient, context.queryKey, context.snapshot);
-    },
-  });
 }
 
 /**
